+++ awc/CHANGES.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.83.
+- Minimum supported Rust version (MSRV) is now 1.88.
+- Fix empty streaming request bodies being sent with chunked transfer encoding.
 
 ## 3.8.1